requirements.txt
absl-py==1.4.0
astor==0.8.1
cached-property==1.5.2
certifi==2021.5.30
charset-normalizer==2.0.12
cycler==0.11.0
dataclasses==0.8
decorator==5.1.1
future==0.18.3
gast==0.5.4
greenlet==2.0.2
grpcio==1.48.2
h5py==2.10.0
idna==3.4
importlib-metadata==4.8.3
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
llvmlite==0.36.0
lxml==4.9.3
Markdown==3.3.7
matplotlib==3.3.4
mock==5.0.2
numba==0.53.1
numpy==1.19.5
obspy==1.2.2
Pillow==8.4.0
protobuf==3.19.6
pyparsing==3.1.1
python-dateutil==2.8.2
PyYAML==6.0
requests==2.27.1
scipy==1.5.4
segyio==1.9.11
six==1.16.0
SQLAlchemy==1.4.49
tensorboard==1.13.1
tensorflow-estimator==1.13.0
tensorflow-gpu==1.13.1
termcolor==1.1.0
typing_extensions==4.1.1
urllib3==1.26.16
Werkzeug==2.0.3
zipp==3.6.0
